backend/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain-community
requests
python-dotenv
PyMuPDF
python-docx
pytesseract
Pillow
streamlit
plotly
fastapi
uvicorn[standard]
python-multipart
livekit
livekit-api
psycopg2
asyncpg
SQLAlchemy
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
bcrypt==4.0.1
beautifulsoup4
urllib3
ollama
# ---- Google Gemini AI ----
google-generativeai==0.8.5

# ---- Document parsing (extra safety for version match) ----
PyPDF2>=3.0.0

# ---- Utility and typing dependencies ----
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
h11==0.16.0
httplib2==0.31.0
idna==3.11
protobuf==5.29.5
pydantic==2.12.4
pydantic_core==2.41.5
pyparsing==3.2.5
sniffio==1.3.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
groq==0.9.0  # Groq API (alternative to Gemini)

# ---- Security Libraries ----
slowapi==0.1.9  # Rate limiting for FastAPI
itsdangerous==2.2.0  # Secure token generation
bleach==6.2.0  # HTML sanitization for XSS protection
python-multipart>=0.0.9  # Required for file uploads (already listed but ensuring version)
email-validator==2.2.0  # Email validation

# ---- AI/ML Libraries ----
accelerate==1.11.0
bitsandbytes==0.48.2
torch==2.9.0
transformers==4.51.0
tokenizers==0.21.4
sentencepiece==0.2.1
safetensors==0.6.2
huggingface-hub==0.36.0
gliner==0.2.22
onnxruntime==1.23.2

# ---- Document Processing ----
annotated-doc==0.0.4
pdfminer.six==20251107
pdfplumber==0.11.8
pypdfium2==5.0.0

# ---- Google API Libraries ----
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2

# ---- Utilities ----
cachetools==6.2.1
cffi==2.0.0
coloredlogs==15.0.1
cryptography==46.0.3
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
httptools==0.7.1
humanfriendly==10.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
psutil==7.1.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rsa==4.9.1
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchfiles==1.1.1
websockets==15.0.1
slowapi
bleach